Rewrite, format, fix shiggy bug

+++ Minecraft.Client/Platform/Common/Audio/stb_vorbis.h
@@ -956,7 +956,8 @@ static int set_file_offset(stb_vorbis *f, unsigned int loc)
    #endif
    f->eof = 0;
    if (USE_MEMORY(f)) {
-      if (f->stream_start + loc >= f->stream_end || f->stream_start + loc < f->stream_start) {
+      if ((size_t)(f->stream_start + loc) >= (size_t)f->stream_end ||
+    (size_t)(f->stream_start + loc) < (size_t)f->stream_start) {
          f->stream = f->stream_end;
          f->eof = 1;
          return 0;
